requirements.txt
tqdm==4.67.*
datasets==3.6.*
torchmetrics==1.7.*
gdown==5.2.*
open_clip_torch==2.32.*
transformers==4.51.*
peft==0.15.*
pandas==2.2.*
scikit-learn==1.6.*
torch==2.6.*